+++ plugins/basic_games/games/game_palworld.py
@@ -11,7 +11,7 @@ from ..basic_features.basic_save_game_info import BasicGameSaveGame
 from ..basic_game import BasicGame
 
 class PalworldGame(BasicGame):
-    Name = "Palworld"
+    Name = "Palworld Support Plugin"
     Author = "WickedSik"
     Version = "0.4.0"
     Description = "Palworld installer with support for multi-platform packages"